Added QuoteService features

+++ springboot/src/main/java/ca/jrvs/apps/trading/model/domain/IexQuote.java
@@ -74,7 +74,7 @@ public class IexQuote {
         this.volume = volume;
     }
 
-    public Long getUpdated() {
+    public Double getUpdated() {
         return updated;
     }
